+++ cmd/opentelemetry-prometheus-sidecar/main.go
@@ -24,7 +24,6 @@ import (
 	"net/url"
 	"os"
 	"runtime"
-	"strings"
 	"time"
 
 	"github.com/go-kit/kit/log"
@@ -39,14 +38,11 @@ import (
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/prometheus"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/retrieval"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/supervisor"
-	"github.com/lightstep/opentelemetry-prometheus-sidecar/tail"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/telemetry"
-	"github.com/oklog/run"
 	"github.com/pkg/errors"
 	"github.com/prometheus/common/version"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/promql/parser"
-	"github.com/prometheus/prometheus/tsdb/wal"
 	grpcMetadata "google.golang.org/grpc/metadata"
 
 	// register grpc compressors
@@ -79,28 +75,6 @@ func main() {
 	}
 }
 
-type prometheusReader interface {
-	Run(context.Context, int) error
-	Next()
-	CurrentSegment() int
-}
-
-func runReader(ctx context.Context, reader prometheusReader, walDir string, startOffset int, maxRetries int) error {
-	var err error
-	attempts := 0
-	for {
-		err = reader.Run(ctx, startOffset)
-		if err != nil && attempts < maxRetries && strings.Contains(err.Error(), tail.ErrSkipSegment.Error()) {
-			_ = retrieval.SaveProgressFile(walDir, startOffset)
-			reader.Next()
-			attempts += 1
-			startOffset = reader.CurrentSegment() * wal.DefaultSegmentSize
-			continue
-		}
-		return err
-	}
-}
-
 func Main() bool {
 	// Setup debugging helpers
 	if os.Getenv("DEBUG") != "" {
@@ -118,21 +92,26 @@ func Main() bool {
 	// Should this process act as supervisor?  This uses the supervisor
 	// environment variable to avoid recursion.
 	isSupervisor := !cfg.DisableSupervisor && os.Getenv(supervisorEnv) == ""
-
-	// Unique identifer for this process.
-	svcInstanceId := uuid.New().String()
-
-	// Configure logging and diagnostics.
 	logger := internal.NewLogger(cfg, isSupervisor)
 
-	telemetry.StaticSetup(logger)
+	scfg := internal.SidecarConfig{
+		ClientFactory:   nil,
+		Monitor:         nil,
+		Logger:          logger,
+		InstanceId:      uuid.New().String(),
+		Matchers:        [][]*labels.Matcher{},
+		MetricRenames:   metricRenames,
+		MetadataCache:   nil,
+		MainConfig:      cfg,
+		FailingReporter: common.NewFailingSet(log.With(logger, "component", "failing_metrics")),
+	}
+
+	telemetry.StaticSetup(scfg.Logger)
 
 	telem := internal.StartTelemetry(
-		cfg,
+		scfg,
 		"opentelemetry-prometheus-sidecar",
-		svcInstanceId,
 		isSupervisor,
-		logger,
 	)
 	if telem != nil {
 		defer telem.Shutdown(context.Background())
@@ -140,15 +119,15 @@ func Main() bool {
 
 	// Start the supervisor.
 	if isSupervisor {
-		return startSupervisor(cfg, telem, logger)
+		return startSupervisor(scfg, telem)
 	}
 
 	// Start the sidecar.  This context lasts the lifetime of the sidecar.
-	ctx, cancelMain := telemetry.ContextWithSIGTERM(logger)
+	ctx, cancelMain := telemetry.ContextWithSIGTERM(scfg.Logger)
 	defer cancelMain()
 
 	healthChecker := health.NewChecker(
-		telem.Controller, cfg.Admin.HealthCheckPeriod.Duration, logger, cfg.Admin.HealthCheckThresholdRatio,
+		telem.Controller, scfg.Admin.HealthCheckPeriod.Duration, scfg.Logger, scfg.Admin.HealthCheckThresholdRatio,
 	)
 
 	httpClient := &http.Client{
@@ -156,167 +135,91 @@ func Main() bool {
 		// Transport: otelhttp.NewTransport(http.DefaultTransport),
 	}
 
-	filters, err := parseFilters(logger, cfg.Filters)
+	scfg.Matchers, err = parseFilters(scfg.Filters)
 	if err != nil {
-		level.Error(logger).Log("msg", "error parsing --filter", "err", err)
+		level.Error(scfg.Logger).Log("msg", "error parsing --filter", "err", err)
 		return false
 	}
 
 	// Parse was validated already, ignore error.
-	promURL, _ := url.Parse(cfg.Prometheus.Endpoint)
+	promURL, _ := url.Parse(scfg.Prometheus.Endpoint)
 
-	promMon := prometheus.NewMonitor(config.PromReady{
-		Logger:                         log.With(logger, "component", "prom_ready"),
+	scfg.Monitor = prometheus.NewMonitor(config.PromReady{
+		Logger:                         log.With(scfg.Logger, "component", "prom_ready"),
 		PromURL:                        promURL,
 		StartupDelayEffectiveStartTime: time.Now(),
 	})
-
-	failingSet := common.NewFailingSet(log.With(logger, "component", "failing"))
 
 	metadataURL, err := promURL.Parse(config.PrometheusMetadataEndpointPath)
 	if err != nil {
 		panic(err)
 	}
-	metadataCache := metadata.NewCache(httpClient, metadataURL, staticMetadata)
+	scfg.MetadataCache = metadata.NewCache(httpClient, metadataURL, staticMetadata)
 
 	// Check the progress file, ensure we can write this file.
-	startOffset, err := readWriteStartOffset(cfg, logger)
+	startOffset, err := readWriteStartOffset(scfg)
 	if err != nil {
-		level.Error(logger).Log("msg", "cannot write progress file", "err", err)
+		level.Error(scfg.Logger).Log("msg", "cannot write progress file", "err", err)
 		return false
 	}
 
-	tailer, err := tail.Tail(
-		ctx,
-		log.With(logger, "component", "wal_reader"),
-		cfg.Prometheus.WAL,
-		promMon,
-	)
+	tailer, err := internal.NewTailer(ctx, scfg)
 	if err != nil {
-		level.Error(logger).Log("msg", "tailing WAL failed", "err", err)
+		level.Error(scfg.Logger).Log("msg", "tailing WAL failed", "err", err)
 		return false
 	}
 
-	outputURL, _ := url.Parse(cfg.Destination.Endpoint)
+	outputURL, _ := url.Parse(scfg.Destination.Endpoint)
 
-	cfg.Destination.Headers[config.AgentKey] = config.AgentMainValue
+	scfg.Destination.Headers[config.AgentKey] = config.AgentMainValue
 
-	scf := internal.NewOTLPClientFactory(otlp.ClientConfig{
-		Logger:           log.With(logger, "component", "storage"),
+	scfg.ClientFactory = internal.NewOTLPClientFactory(otlp.ClientConfig{
+		Logger:           log.With(scfg.Logger, "component", "storage"),
 		URL:              outputURL,
-		Timeout:          cfg.Destination.Timeout.Duration,
-		RootCertificates: cfg.Security.RootCertificates,
-		Headers:          grpcMetadata.New(cfg.Destination.Headers),
-		Compressor:       cfg.Destination.Compression,
-		Prometheus:       cfg.Prometheus,
-		FailingSet:       failingSet,
+		Timeout:          scfg.Destination.Timeout.Duration,
+		RootCertificates: scfg.Security.RootCertificates,
+		Headers:          grpcMetadata.New(scfg.Destination.Headers),
+		Compressor:       scfg.Destination.Compression,
+		Prometheus:       scfg.Prometheus,
+		FailingReporter:  scfg.FailingReporter,
 	})
-
-	queueManager, err := otlp.NewQueueManager(
-		log.With(logger, "component", "queue_manager"),
-		cfg.QueueConfig(),
-		cfg.Destination.Timeout.Duration,
-		scf,
-		tailer,
-		retrieval.LabelsToResource(createPrimaryDestinationResourceLabels(svcInstanceId, cfg.Destination.Attributes)),
-	)
-	if err != nil {
-		level.Error(logger).Log("msg", "creating queue manager failed", "err", err)
-		return false
-	}
 
 	// Start the admin server.
 	go func() {
 		defer cancelMain()
 
-		server := newAdminServer(healthChecker, cfg.Admin, logger)
+		server := newAdminServer(healthChecker, scfg.Admin, scfg.Logger)
 
 		go func() {
-			level.Debug(logger).Log("msg", "starting admin server")
+			level.Debug(scfg.Logger).Log("msg", "starting admin server")
 			<-ctx.Done()
 			if err := server.Shutdown(context.Background()); err != nil {
-				level.Error(logger).Log("msg", "admin server shutdown", "err", err)
+				level.Error(scfg.Logger).Log("msg", "admin server shutdown", "err", err)
 			}
 		}()
 
 		if err := server.ListenAndServe(); err != http.ErrServerClosed {
-			level.Error(logger).Log("msg", "admin listener", "err", err)
+			level.Error(scfg.Logger).Log("msg", "admin listener", "err", err)
 		}
 	}()
 
-	logStartup(cfg, logger)
+	logStartup(cfg, scfg.Logger)
 
 	// Test for Prometheus and Outbound dependencies before starting.
-	if err := selfTest(ctx, scf, cfg.StartupTimeout.Duration, logger, promMon); err != nil {
-		level.Error(logger).Log("msg", "selftest failed, not starting", "err", err)
+	if err := selfTest(ctx, scfg); err != nil {
+		level.Error(scfg.Logger).Log("msg", "selftest failed, not starting", "err", err)
 		return false
 	}
 
-	level.Debug(logger).Log("msg", "entering run state")
+	level.Debug(scfg.Logger).Log("msg", "entering run state")
 	healthChecker.SetRunning()
 
-	prometheusReader := retrieval.NewPrometheusReader(
-		log.With(logger, "component", "prom_wal"),
-		cfg.Prometheus.WAL,
-		tailer,
-		filters,
-		metricRenames,
-		metadataCache,
-		queueManager,
-		cfg.OpenTelemetry.MetricsPrefix,
-		cfg.Prometheus.MaxPointAge.Duration,
-		promMon.GetScrapeConfig(),
-	)
-
-	// Run two inter-depdendent components:
-	// (1) Prometheus reader
-	// (2) Queue manager
-	// TODO: Replace this with x/sync/errgroup
-	var g run.Group
-	{
-		g.Add(
-			func() error {
-				level.Info(logger).Log("msg", "starting Prometheus reader", "segment", startOffset/wal.DefaultSegmentSize)
-				return runReader(ctx, prometheusReader, cfg.Prometheus.WAL, startOffset, config.DefaultMaxRetrySkipSegments)
-			},
-			func(err error) {
-				// Prometheus reader needs to be stopped before closing the TSDB
-				// so that it doesn't try to write samples to a closed storage.
-				// See the use of `stopCh` below to explain how this works.
-				level.Info(logger).Log("msg", "stopping Prometheus reader")
-				cancelMain()
-			},
-		)
-	}
-	{
-		stopCh := make(chan struct{})
-		g.Add(
-			func() error {
-				if err := queueManager.Start(); err != nil {
-					return err
-				}
-				level.Info(logger).Log("msg", "starting OpenTelemetry writer")
-				<-stopCh
-				return nil
-			},
-			func(err error) {
-				if err := queueManager.Stop(); err != nil {
-					level.Error(logger).Log(
-						"msg", "stopping OpenTelemetry writer",
-						"err", err,
-					)
-				}
-				close(stopCh)
-			},
-		)
-	}
-	if err := g.Run(); err != nil {
-		level.Error(logger).Log("msg", "run loop error", "err", err)
-		return false
+	if err = internal.StartComponents(ctx, scfg, tailer, startOffset); err != nil {
+		cancelMain()
 	}
 
 	// SIGTERM causes graceful shutdown.
-	level.Info(logger).Log("msg", "sidecar process exiting")
+	level.Info(scfg.Logger).Log("msg", "sidecar process exiting")
 	return true
 }
 
@@ -331,7 +234,7 @@ func usage(err error) {
 
 // parseFilters parses two flags that contain PromQL-style metric/label selectors and
 // returns a list of the resulting matchers.
-func parseFilters(logger log.Logger, filters []string) ([][]*labels.Matcher, error) {
+func parseFilters(filters []string) ([][]*labels.Matcher, error) {
 	var matchers [][]*labels.Matcher
 	for _, f := range filters {
 		m, err := parser.ParseMetricSelector(f)
@@ -343,35 +246,22 @@ func parseFilters(logger log.Logger, filters []string) ([][]*labels.Matcher, err
 	return matchers, nil
 }
 
-// createPrimaryDestinationResourceLabels returns the OTLP resources
-// to use for the primary destination.
-func createPrimaryDestinationResourceLabels(svcInstanceId string, extraLabels map[string]string) labels.Labels {
-	// Note: there is minor benefit in including an external label
-	// to indicate the process ID here.  See
-	// https://github.com/lightstep/opentelemetry-prometheus-sidecar/issues/44
-	// Until resources are serialized once per request, leave this
-	// commented out (and a test in e2e_test.go):
-	// extraLabels[externalLabelPrefix+string(semconv.ServiceInstanceIDKey)]
-	// = svcInstanceId
-	return labels.FromMap(extraLabels)
-}
+func selfTest(ctx context.Context, scfg internal.SidecarConfig) error {
+	client := scfg.ClientFactory.New()
 
-func selfTest(ctx context.Context, scf otlp.StorageClientFactory, timeout time.Duration, logger log.Logger, promMon *prometheus.Monitor) error {
-	client := scf.New()
-
-	ctx, cancel := context.WithTimeout(ctx, timeout)
+	ctx, cancel := context.WithTimeout(ctx, scfg.StartupTimeout.Duration)
 	defer cancel()
 
-	level.Debug(logger).Log("msg", "checking Prometheus readiness")
+	level.Debug(scfg.Logger).Log("msg", "checking Prometheus readiness")
 
 	// These tests are performed sequentially, to keep the logs simple.
 	// Note WaitForReady loops until success or stop if the context is canceled
 	// or an unsupported version of prometheus is identified
-	if err := promMon.WaitForReady(ctx, cancel); err != nil {
+	if err := scfg.Monitor.WaitForReady(ctx, cancel); err != nil {
 		return errors.Wrap(err, "Prometheus is not ready")
 	}
 
-	level.Debug(logger).Log("msg", "checking OpenTelemetry endpoint")
+	level.Debug(scfg.Logger).Log("msg", "checking OpenTelemetry endpoint")
 
 	// Outbound connection test.
 	{
@@ -385,7 +275,7 @@ func selfTest(ctx context.Context, scf otlp.StorageClientFactory, timeout time.D
 		}
 	}
 
-	level.Debug(logger).Log("msg", "selftest was successful")
+	level.Debug(scfg.Logger).Log("msg", "selftest was successful")
 	return nil
 }
 
@@ -407,10 +297,10 @@ func logStartup(cfg config.MainConfig, logger log.Logger) {
 	}
 }
 
-func startSupervisor(cfg config.MainConfig, telem *telemetry.Telemetry, logger log.Logger) bool {
+func startSupervisor(scfg internal.SidecarConfig, telem *telemetry.Telemetry) bool {
 	super := supervisor.New(supervisor.Config{
-		Logger:    logger,
-		Admin:     cfg.Admin,
+		Logger:    scfg.Logger,
+		Admin:     scfg.Admin,
 		Telemetry: telem,
 	})
 
@@ -433,13 +323,13 @@ func newAdminServer(hc *health.Checker, acfg config.AdminConfig, logger log.Logg
 
 // readWriteStartOffset reads the last (approxiate) progress position and re-writes
 // the progress file, to ensure we have write permission on startup.
-func readWriteStartOffset(cfg config.MainConfig, logger log.Logger) (int, error) {
-	startOffset, err := retrieval.ReadProgressFile(cfg.Prometheus.WAL)
+func readWriteStartOffset(scfg internal.SidecarConfig) (int, error) {
+	startOffset, err := retrieval.ReadProgressFile(scfg.Prometheus.WAL)
 	if err != nil {
-		level.Warn(logger).Log("msg", "reading progress file failed", "err", err)
+		level.Warn(scfg.Logger).Log("msg", "reading progress file failed", "err", err)
 		startOffset = 0
 	}
 
-	err = retrieval.SaveProgressFile(cfg.Prometheus.WAL, startOffset)
+	err = retrieval.SaveProgressFile(scfg.Prometheus.WAL, startOffset)
 	return startOffset, err
 }