+++ otlp/client.go
@@ -95,7 +95,7 @@ type Client struct {
 	headers          grpcMetadata.MD
 	compressor       string
 	prometheus       config.PromConfig
-	invalidSet       *common.FailingSet
+	invalidSet       common.FailingReporter
 
 	conn *grpc.ClientConn
 }
@@ -109,7 +109,7 @@ type ClientConfig struct {
 	Headers          grpcMetadata.MD
 	Compressor       string
 	Prometheus       config.PromConfig
-	FailingSet       *common.FailingSet
+	FailingReporter  common.FailingReporter
 }
 
 // NewClient creates a new Client.
@@ -126,7 +126,7 @@ func NewClient(conf ClientConfig) *Client {
 		headers:          conf.Headers,
 		compressor:       conf.Compressor,
 		prometheus:       conf.Prometheus,
-		invalidSet:       conf.FailingSet,
+		invalidSet:       conf.FailingReporter,
 	}
 }